+++ src/navigations/index.tsx
@@ -43,13 +43,6 @@ import { AppState } from '@store/types';
 import { AuthInfo, UserState } from '@store/User/UserTypes';
 import { navigationRef, navigate } from '@notifications';
 import { Notif } from '@store/Notif/NotifTypes';
-import {
-  SupportFAQTypes,
-  DeliveryReportTypes,
-  Category,
-  Image,
-  MailTypes,
-} from 'types';
 import Topbar, {
   setTitle,
   topbarRef,
@@ -60,99 +53,9 @@ import { NavigationContainer } from '@react-navigation/native';
 import { Platform } from 'react-native';
 import { WINDOW_WIDTH, WINDOW_HEIGHT } from '@utils';
 import i18n from '@i18n';
+import { Screens, AuthStackParamList, AppStackParamList } from '@utils/Screens';
 
 export { navigationRef, navigate };
-
-export enum Screens {
-  Begin = 'Begin',
-  Splash = 'Splash',
-  Login = 'Login',
-  Terms = 'Terms',
-  Privacy = 'Privacy',
-  RegisterCreds = 'RegisterCreds',
-  RegisterPersonal = 'RegisterPersonal',
-  RegisterAddress = 'RegisterAddress',
-  ReferFriends = 'ReferFriends',
-  Delivery = 'DeliveryReporting',
-  ContactSelector = 'ContactSelector',
-  ChooseCategory = 'ChooseCategory',
-  ChooseOption = 'ChooseOption',
-  ComposeLetter = 'ComposeLetter',
-  ComposePostcard = 'ComposePostcard',
-  ReviewLetter = 'ReviewLetter',
-  ReviewPostcard = 'ReviewPostcard',
-  ContactInfo = 'ContactInfo',
-  FacilityDirectory = 'FacilityDirectory',
-  AddManually = 'AddManually',
-  ReviewContact = 'ReviewContact',
-  Issues = 'Issues',
-  IssuesDetail = 'IssuesDetail',
-  IssuesDetailSecondary = 'IssuesDetailSecondary',
-  SingleContact = 'SingleContact',
-  MailTracking = 'MailTracking',
-  MemoryLane = 'MemoryLane',
-  MailDetails = 'MailDetails',
-  SupportFAQ = 'SupportFAQ',
-  SupportFAQDetail = 'SupportFAQDetail',
-  UpdateContact = 'UpdateContact',
-  UpdateProfile = 'UpdateProfile',
-}
-
-export type AuthStackParamList = {
-  Splash: undefined;
-  Begin: undefined;
-  Login: undefined;
-  Terms: undefined;
-  Privacy: undefined;
-  RegisterCreds: undefined;
-  RegisterPersonal: {
-    email: string;
-    password: string;
-    passwordConfirmation: string;
-    remember: boolean;
-  };
-  RegisterAddress: {
-    email: string;
-    password: string;
-    passwordConfirmation: string;
-    remember: boolean;
-    firstName: string;
-    lastName: string;
-    referrer: string;
-    image: Image | undefined;
-  };
-};
-
-export type AppStackParamList = {
-  AddManually: { phyState: string };
-  ChooseCategory: undefined;
-  ChooseOption: undefined;
-  ComposeLetter: undefined;
-  ComposePostcard: { category: Category };
-  ContactInfo: { addFromSelector?: boolean; phyState?: string };
-  ContactSelector: undefined;
-  FacilityDirectory: { phyState: string };
-  InmateLocator: { uri: string };
-  Issues: undefined;
-  IssuesDetail: { issue: DeliveryReportTypes } | undefined;
-  IssuesDetailSecondary: { issue: DeliveryReportTypes } | undefined;
-  LetterPreview: undefined;
-  PostcardPreview: undefined;
-  MailDetails: undefined;
-  MailTracking: undefined;
-  MemoryLane: undefined;
-  ReferFriends: { mailType: MailTypes };
-  ReviewLetter: undefined;
-  ReviewPostcard: { horizontal: boolean; category: string };
-  ReviewContact: { manual: boolean };
-  Setup: undefined;
-  SingleContact: undefined;
-  Splash: undefined;
-  SupportFAQ: undefined;
-  SupportFAQDetail: { issue: SupportFAQTypes } | undefined;
-  UpdateContact: { contactId: number } | undefined;
-  UpdateProfile: undefined;
-};
 
 interface RouteDetails {
   title: string;
@@ -320,7 +223,7 @@ const NavigatorBase: React.FC<Props> = (props: Props) => {
           options={{ cardStyleInterpolator: fadeTransition }}
         />
         <Stack.Screen
-          name={Screens.RegisterAddress}
+          name={Screens.ChooseCategory}
           component={ChooseCategoryScreen}
         />
         <Stack.Screen
