+++ src/store/index.ts
@@ -4,6 +4,7 @@ import { AsyncStorage } from 'react-native';
 import { AppState } from './types';
 import CategoryReducer from './Category/CategoryReducer';
 import ContactReducer from './Contact/ContactReducer';
+import FacilityReducer from './Facility/FacilityReducer';
 import MailReducer from './Mail/MailReducer';
 import NotifReducer from './Notif/NotifReducer';
 import UserReducer from './User/UserReducer';
@@ -11,7 +12,7 @@ import UserReducer from './User/UserReducer';
 const config = {
   key: 'root',
   storage: AsyncStorage,
-  blacklist: ['user'],
+  blacklist: ['user', 'facility'],
   whitelist: ['category', 'contact', 'notif', 'mail'],
 };
 
@@ -19,6 +20,7 @@ const combinedReducers = combineReducers<AppState>({
   user: UserReducer,
   category: CategoryReducer,
   contact: ContactReducer,
+  facility: FacilityReducer,
   notif: NotifReducer,
   mail: MailReducer,
 });
