+++ src/store/types.ts
@@ -1,5 +1,6 @@
 import { CategoryState } from './Category/CategoryTypes';
 import { ContactState } from './Contact/ContactTypes';
+import { FacilityState } from './Facility/FacilityTypes';
 import { MailState } from './Mail/MailTypes';
 import { NotifState } from './Notif/NotifTypes';
 import { UserState } from './User/UserTypes';
@@ -7,6 +8,7 @@ import { UserState } from './User/UserTypes';
 export interface AppState {
   category: CategoryState;
   contact: ContactState;
+  facility: FacilityState;
   mail: MailState;
   notif: NotifState;
   user: UserState;
