+++ src/utils/index.ts
@@ -9,7 +9,6 @@ import i18n from '@i18n';
 import * as Segment from 'expo-analytics-segment';
 import { addBusinessDays } from 'date-fns';
 import Constants from 'expo-constants';
-import { Screens } from '@navigations';
 import {
   ABBREV_TO_STATE,
   STATE_TO_ABBREV,
@@ -18,6 +17,7 @@ import {
 } from './States';
 import { Prompts, getRandomPromptIx } from './FeelingStuck';
 import REFERERS from './Referers';
+import { Screens } from './Screens';
 
 export { Prompts, getRandomPromptIx };
 export { REFERERS };
