+++ src/views/AddContact/ContactInfo.react.tsx
@@ -8,7 +8,7 @@ import {
 } from 'react-native';
 import { Button, Icon, Input, KeyboardAvoider } from '@components';
 import { Colors, Typography } from '@styles';
-import { AppStackParamList, Screens } from '@navigations';
+import { AppStackParamList, Screens } from '@utils/Screens';
 import { StackNavigationProp } from '@react-navigation/stack';
 import {
   LOB_NAME_CHAR_LIMIT,
@@ -26,6 +26,7 @@ import Letter from '@assets/views/AddContact/Letter';
 import { setProfileOverride } from '@components/Topbar/Topbar.react';
 import * as Segment from 'expo-analytics-segment';
 import { ContactPersonal, ContactDraft } from 'types';
+import { getFacilities } from '@api';
 import CommonStyles from './AddContact.styles';
 
 type ContactInfoScreenNavigationProp = StackNavigationProp<
@@ -343,10 +344,14 @@ class ContactInfoScreenBase extends React.Component<Props, State> {
                     if (
                       this.stateRef.current &&
                       this.stateRef.current.state.valid
-                    )
+                    ) {
+                      const abbrev =
+                        STATE_TO_ABBREV[this.stateRef.current.state.value];
+                      if (abbrev) getFacilities(abbrev);
                       this.setState({
-                        stateToSearch: this.stateRef.current?.state.value,
+                        stateToSearch: this.stateRef.current.state.value,
                       });
+                    }
                     this.updateValid();
                   }}
                   onInvalid={() => this.setState({ valid: false })}
