+++ src/views/Compose/ComposeLetter.react.tsx
@@ -15,7 +15,7 @@ import {
   KeyboardAvoider,
 } from '@components';
 import { StackNavigationProp } from '@react-navigation/stack';
-import { AppStackParamList, Screens } from '@navigations';
+import { AppStackParamList, Screens } from '@utils/Screens';
 import { connect } from 'react-redux';
 import { AppState } from '@store/types';
 import { setContent, setImage } from '@store/Mail/MailActions';
