+++ src/views/Compose/ReviewPostcard.react.tsx
@@ -2,7 +2,7 @@ import React, { Dispatch } from 'react';
 import { TouchableOpacity, View, Keyboard, Text } from 'react-native';
 import { StaticPostcard } from '@components';
 import { StackNavigationProp } from '@react-navigation/stack';
-import { AppStackParamList, Screens } from '@navigations';
+import { AppStackParamList, Screens } from '@utils/Screens';
 import { Draft, Contact, MailTypes } from 'types';
 import { AppState } from '@store/types';
 import { connect } from 'react-redux';
