+++ src/views/MemoryLane/MailDetails.react.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { ScrollView, Text, View, Image } from 'react-native';
-import { AppStackParamList } from '@navigations';
+import { AppStackParamList } from '@utils/Screens';
 import { StackNavigationProp } from '@react-navigation/stack';
 import { Mail, MailTypes, MailStatus } from 'types';
 import { connect } from 'react-redux';
