+++ src/views/Onboarding/RegisterAddress.react.tsx
@@ -7,7 +7,7 @@ import {
   View,
 } from 'react-native';
 import { StackNavigationProp } from '@react-navigation/stack';
-import { AuthStackParamList, Screens } from '@navigations';
+import { AuthStackParamList, Screens } from '@utils/Screens';
 import { Input, KeyboardAvoider } from '@components';
 import i18n from '@i18n';
 import { Typography } from '@styles';
