+++ src/views/Onboarding/RegisterPersonal.react.tsx
@@ -1,7 +1,7 @@
 import React, { createRef } from 'react';
 import { ScrollView, Text, TouchableOpacity, View } from 'react-native';
 import { StackNavigationProp } from '@react-navigation/stack';
-import { AuthStackParamList, Screens } from '@navigations';
+import { AuthStackParamList, Screens } from '@utils/Screens';
 import { Input, PicUpload, KeyboardAvoider } from '@components';
 import i18n from '@i18n';
 import { Typography } from '@styles';
