+++ src/views/Report/IssuesDetail.react.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Text, View, TextStyle, ViewStyle } from 'react-native';
 import { Typography } from '@styles';
 import { Button, Icon } from '@components';
-import { AppStackParamList, Screens } from '@navigations';
+import { AppStackParamList, Screens } from '@utils/Screens';
 import { StackNavigationProp } from '@react-navigation/stack';
 import i18n from '@i18n';
 import { DeliveryReportTypes } from 'types';
