+++ src/views/Report/SupportFAQ.react.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Text, View, ScrollView } from 'react-native';
 import { Typography } from '@styles';
 import { ListItem } from '@components';
-import { AppStackParamList } from '@navigations';
+import { AppStackParamList } from '@utils/Screens';
 import { StackNavigationProp } from '@react-navigation/stack';
 import i18n from '@i18n';
 import { SupportFAQTypes } from 'types';
